removed unnessery file and function
