+++ packages/excalidraw/element/binding.ts
@@ -1212,12 +1212,18 @@ export const fixBindingsAfterDuplication = (
 ): void => {
   // First collect all the binding/bindable elements, so we only update
   // each once, regardless of whether they were duplicated or not.
-  const allBoundElementIds: Set<ExcalidrawElement["id"]> = new Set();
-  const allBindableElementIds: Set<ExcalidrawElement["id"]> = new Set();
+  const allBoundElementIds: Set<ExcalidrawElement["id"]> = new Set(); // arrow and linearElement
+  const allBindableElementIds: Set<ExcalidrawElement["id"]> = new Set(); // shapes like rectange and squares and dimond and circle
+
+  // old shapes collection with types
+  const linearShapesIds: Set<ExcalidrawElement["id"]> = new Set();
+  const basicShapesIds: Set<ExcalidrawElement["id"]> = new Set();
+
   const shouldReverseRoles = duplicatesServeAsOld === "duplicatesServeAsOld";
   const duplicateIdToOldId = new Map(
     [...oldIdToDuplicatedId].map(([key, value]) => [value, key]),
   );
+
   oldElements.forEach((oldElement) => {
     const { boundElements } = oldElement;
     if (boundElements != null && boundElements.length > 0) {
@@ -1226,6 +1232,7 @@ export const fixBindingsAfterDuplication = (
           allBoundElementIds.add(boundElement.id);
         }
       });
+      basicShapesIds.add(oldElement.id); // this is for geting the old shapes from all shapes and then updated them
       allBindableElementIds.add(oldIdToDuplicatedId.get(oldElement.id)!);
     }
     if (isBindingElement(oldElement)) {
@@ -1243,6 +1250,7 @@ export const fixBindingsAfterDuplication = (
       }
       if (oldElement.startBinding != null || oldElement.endBinding != null) {
         allBoundElementIds.add(oldIdToDuplicatedId.get(oldElement.id)!);
+        linearShapesIds.add(oldElement.id); // we have to update the old linear element as well
       }
     }
   });
@@ -1263,6 +1271,86 @@ export const fixBindingsAfterDuplication = (
     });
   });
 
+  // update the origial linear element
+
+  (
+    sceneElements.filter(({ id }) =>
+      linearShapesIds.has(id),
+    ) as ExcalidrawLinearElement[]
+  ).forEach((element) => {
+    const { startBinding, endBinding } = element;
+    // if startBinding and endBinding ids exsist in the selected list (oldIdToDuplicatedId) then we are not going to change else change
+    // because if that in the list then we know it's going to get duplicated and the duplicate one is going to get connected to the duplicate arrow
+    // so no need to change
+    // Track if we need to mutate the element
+    let updatedStartBinding = startBinding;
+    let updatedEndBinding = endBinding;
+
+    // Check if startBinding needs to be changed
+    if (
+      startBinding != null &&
+      !oldIdToDuplicatedId.has(startBinding.elementId)
+    ) {
+      updatedStartBinding = null;
+      //
+      sceneElements
+        .filter((ele) => ele.id === startBinding.elementId)
+        .forEach((ele) => {
+          const newBoundElements = ele.boundElements?.map((boundEle) => {
+            if (boundEle.id === element.id) {
+              return {
+                id: oldIdToDuplicatedId.get(element.id),
+                type: element.type,
+              } as Readonly<{
+                id: ExcalidrawLinearElement["id"];
+                type: "arrow" | "text";
+              }>;
+            }
+            return boundEle;
+          });
+          mutateElement(ele, {
+            boundElements: newBoundElements,
+          });
+        });
+    }
+
+    // Check if endBinding needs to be changed
+    if (endBinding != null && !oldIdToDuplicatedId.has(endBinding.elementId)) {
+      updatedEndBinding = null;
+
+      sceneElements
+        .filter((ele) => ele.id === endBinding.elementId)
+        .forEach((ele) => {
+          const newBoundElements = ele.boundElements?.map((boundEle) => {
+            if (boundEle.id === element.id) {
+              return {
+                id: oldIdToDuplicatedId.get(element.id),
+                type: element.type,
+              } as Readonly<{
+                id: ExcalidrawLinearElement["id"];
+                type: "arrow" | "text";
+              }>;
+            }
+            return boundEle;
+          });
+          mutateElement(ele, {
+            boundElements: newBoundElements,
+          });
+        });
+    }
+
+    // Mutate only if there are changes to either binding
+    if (
+      updatedStartBinding !== startBinding ||
+      updatedEndBinding !== endBinding
+    ) {
+      mutateElement(element, {
+        startBinding: updatedStartBinding,
+        endBinding: updatedEndBinding,
+      });
+    }
+  });
+
   // Update the bindable shapes
   sceneElements
     .filter(({ id }) => allBindableElementIds.has(id))
@@ -1273,6 +1361,7 @@ export const fixBindingsAfterDuplication = (
       )?.boundElements;
 
       if (boundElements && boundElements.length > 0) {
+        // i gess we are changing the new duplicated one here but the issue is that we have to change the original as well
         mutateElement(bindableElement, {
           boundElements: boundElements.map((boundElement) =>
             oldIdToDuplicatedId.has(boundElement.id)
@@ -1282,6 +1371,23 @@ export const fixBindingsAfterDuplication = (
                 }
               : boundElement,
           ),
+        });
+      }
+    });
+
+  // change the orignal basic_shapes as well
+  sceneElements
+    .filter(({ id }) => {
+      return basicShapesIds.has(id);
+    })
+    .forEach((element) => {
+      const { boundElements } = element;
+      if (boundElements != null && boundElements.length > 0) {
+        const newBoundElements = boundElements.filter(
+          (ele) => oldIdToDuplicatedId.has(ele.id), // if the element is also get duplicated with the shapes then we want only that one else not
+        );
+        mutateElement(element, {
+          boundElements: newBoundElements,
         });
       }
     });
